v1.5.9

+++ OSCVRCWiz/Services/Integrations/Heartrate/HeartratePulsoid.cs
@@ -217,8 +217,8 @@ namespace OSCVRCWiz.Services.Integrations.Heartrate
             var message3 = new CoreOSC.OscMessage("/avatar/parameters/hundredsHR", hundreds);
             OSC.OSCSender.Send(message3);
 
-            float HRPercent = currentHR / 256;
-            var message4 = new CoreOSC.OscMessage("/avatar/parameters/HRPercent", HRPercent);
+            float HRPercent = (float)currentHR / 256;
+            var message4 = new CoreOSC.OscMessage("/avatar/parameters/HRPercent", (float)HRPercent);
             OSC.OSCSender.Send(message4);
